Code climate: backend/sdoc: relocate Reference, merge type_system into models
This simplifies tracing the L2 requirement of SDoc model.
