+++ tests/unit/strictdoc/backend/sdoc/test_requirement_from_dict.py
@@ -1,6 +1,6 @@
 from strictdoc.backend.sdoc.models.document_grammar import DocumentGrammar
+from strictdoc.backend.sdoc.models.model import RequirementFieldName
 from strictdoc.backend.sdoc.models.object_factory import SDocObjectFactory
-from strictdoc.backend.sdoc.models.type_system import RequirementFieldName
 
 
 def test_010_full_dict():
